requirements.txt
mysql-connector==2.2.9
Pillow==7.1.2
pycoingecko==1.2.0
pyTelegramBotAPI==3.7.1
requests==2.7.0
schedule==0.6.0
six==1.14.0
supervisor==4.2.0
telebot==0.0.3
